#18 create development environment
