+++ bin/app
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+node ./src/app.js
