+++ nginx.conf
@@ -8,40 +8,64 @@ events {
 }
 
 http {
-	upstream api {
-		server 127.0.0.1:8080;
-	}
+	root /home/kikurage/server/static;
 
 	log_format with_time '$remote_addr - $remote_user [$time_local] '
 						 '"$request" $status $body_bytes_sent '
 						 '"$http_referer" "$http_user_agent" $request_time';
+
 	access_log /var/log/nginx/access.log with_time;
 	include /etc/nginx/mime.types;
 
-	server {
-		root /home/kikurage/server/static;
+	upstream production {
+		server 127.0.0.1:8081;
+	}
 
-		#listen 443 default ssl;
-	    #ssl on;
-		#ssl_certificate      /home/kikurage/server/ssl/nginx.pem;
-		#ssl_certificate_key  /home/kikurage/server/ssl/nginx.key;
+	upstream dev {
+		server 127.0.0.1:8080;
+	}
+
+	server {
+		server_name atami.kikurage.xyz;
 
 		location /bower_components {
 			root /home/kikurage/server/;
-			try_files $uri @api;
+			try_files $uri 404;
 		}
 
 		location / {
 			add_header    X-Real-IP       $remote_addr;
 			add_header    X-Forwarded-For $proxy_add_x_forwarded_for;
 			add_header    Host            $http_host;
-	        try_files $uri @api;
+	        try_files $uri @production;
 	    }
 
-		location @api {
+		location @production {
 			add_header Access-Control-Allow-Origin *;
 			proxy_set_header Host $host;
-			proxy_pass http://api;
+			proxy_pass http://production;
+		}
+	}
+
+	server {
+		server_name atami-dev.kikurage.xyz;
+
+		location /bower_components {
+			root /home/kikurage/server/;
+			try_files $uri 404;
+		}
+
+		location / {
+			add_header    X-Real-IP       $remote_addr;
+			add_header    X-Forwarded-For $proxy_add_x_forwarded_for;
+			add_header    Host            $http_host;
+	        try_files $uri @dev;
+	    }
+
+		location @dev {
+			add_header Access-Control-Allow-Origin *;
+			proxy_set_header Host $host;
+			proxy_pass http://dev;
 		}
 	}
 }
