+++ src/config/dev.json
@@ -1,0 +1,9 @@
+{
+	"name": "dev",
+	"server": {
+		"port": "8080"
+	},
+	"db": {
+		"url": "mongodb://localhost:27017/atami-dev"
+	}
+}
