--- upload
@@ -1,1 +1,0 @@
-scp -r ./app ./static ./nginx.conf ./bower.json ./ssl ./package.json install stoguVPS:~/server/
